webservice/requirements.txt
certifi==2017.4.17
chardet==3.0.4
Django==1.11.3
django-cors-headers==2.1.0
django-filter==1.0.4
djangorestframework==3.6.3
idna==2.5
Markdown==2.6.8
numpy==1.13.1
olefile==0.44
Pillow==4.2.1
pytz==2017.2
requests==2.18.1
urllib3==1.21.1
